+++ src/en/kunmanga/build.gradle
@@ -2,8 +2,8 @@ ext {
     extName = 'Kun Manga'
     extClass = '.KunManga'
     themePkg = 'madara'
-    baseUrl = 'https://www.kunmanga.online'
-    overrideVersionCode = 2
+    baseUrl = 'https://kunmanga.com'
+    overrideVersionCode = 1
     isNsfw = true
 }
